feat: restart local tracks, improved video mute
