+++ example/index.html
@@ -98,15 +98,14 @@
                 >
                   Mute Video
                 </button>
-
                 <button
-                  id="toggle-video-button"
+                  id="flip-video-button"
                   class="btn btn-secondary"
                   disabled
                   type="button"
-                  onclick="toggleVideo()"
+                  onclick="flipVideo()"
                 >
-                  Toggle Video
+                  Flip Video
                 </button>
                 <button
                   id="share-screen-button"
